Separate types, fix createPost function
I wanted to try optimizing the whole function chain
when creating a post. First, it needs to create the
post itself, sending the necessary data for the
INSERT query. Afterwards, if there are tags present,
it needs to create a new tag for each (up to 3). Lastly,
it needs the post_id and tag_id, both returned from their
respective queries. Apparently, simply looping through them
is a terrible idea according to the `pg-promise` FAQ. And so,
initially, I wanted to use the task method and the tx method
together. However, it keeps on giving me errors. What I wanted
was to go through the whole chain in one request:
INSERT -> SELECT -> INSERT
Unfortunately, it kept on spitting out BatchErrors/DatabaseErrors:
"current transaction is aborted, commands ignored etc. etc."
which apparently means that there was a fuckup along the way
so it decided to shut itself down. I didn't know how to fix it;
it probably had to do something with race conditions or something.
It probably would have worked if I tried turning it into a single
query like what the author said in the FAQ but I couldn't figure it out.
I was able to create (aka copy an answer) an INSERT -> SELECT
query but an INSERT -> SELECT -> INSERT query was too much for me.
So I had to make do with 3 connections in a single function call:
1. creating the post (returning post id)
2. creating/checking the tags* (returning tag_id)
3. creating the post/tag relation
*The tags are in one tx so that's something positive at least
It's most likely some garbage optimization but at least it works;
the tests aren't shitting on me anymore, at least for now.