Change formatting, remove catchall routes
I created an array of routes that I want to go through a specific
function first, that is, authentication. Either checking of the
password first, or of the request's authentication token, or whatever
middleware I had in mind. I removed the catchall routes, changing it
to what I have now because I should make it go through validation
first before doing any sort of verification against the database.