Redo tests due to change in authToken
Fix the tests to accomodate changes in authToken's verifyToken function;
I added a check to deny requests if the username in the decoded token
doesn't exist in the database. For some reason, even if the request
doesn't go to that branch where it would get denied, there's still new
behaviour with the tests so I had to fix it. It talked about errors
about setting headers after it has been sent to the client. I managed to
fix it by reinitializing the agent each time so that it doesn't act as
the same person/entity when sending the requests. It's better like this
regardless so that the tests I wrote are independent instead of what it
was previously. Still, I'm not sure what's with the errors. At least
it's ""fixed""...