Add followStatus to returned value
Add a query to check for the followStatus between two accounts (if the
loggedInUsername is provided, else just do previous behaviour). The
return value is then the profile info (same as before) and the follow
status. This reduces the amount of times the database is queried when
visiting a profile. Previously, after getting the profile info, another
separate query was made just to confirm whether or not the visited
profile is already followed which affects the display and behaviour of
the follow button in the profile page.