Consolidate functions
Combine the functions "checkIfFollowed", "followUser", and
"unfollowUser" into one "toggleFollow" function. Previously, the check
endpoint was used to determine what was to be display in the profile
page. The check was implemented in the getProfileInfo function in the
profile model to reduce the trips to the database. Likewise, instead of
checking the follow status to determine if the toggle was to be a
"follow" or an "unfollow", the check was just implemented in a function
combined with the toggle. Fixed the tests and removed unnecessary
endpoints (the check) accordingly