Add check for if username exists
Just readding the previously removed function with changes in parsing
the request (from POST to GET). Currently, with the getProfileInfo
function, it will return true or false if the logged in user is
following the visited profile. It has no checks whether or not the
profile actually exists. It will just return false if so, there aren't
any null checks. An authentication middleware is necessary.