Fix functions
Fix createPost to accept tags with array type as input. Any string
manipulation is now unnecesary because of the change in data type and
because the validation of the tags input is now done in the client.

The getPosts and searchPosts function also now return 11 posts instead
of 10 as a tweak to help with pagination, for example:
If the returned count is only 10, it means that there is no other posts
remaining, that is, it is the last page already. Else, there is still at
least one, meaning the next page button should not be disabled.