Standardize form behaviour
I added a function disable the button upon submission, waiting for a
response before reenabling it. This is to prevent multiple submissions
when the request is still processing. The response after submission and
the behaviour after receiving a response have also been changed to make
it all similar (as much as possible, at least). Lastly, the variable
names have also been changed to make it all the same:
req, onSubmit, buttonState
because trying to fix things was starting to become a pain