Fix form validation
Missed the case where a "space" character would get through. I also had
to be clear with the difference between validation for optional and
required tags. Lastly, I had to be explicit with the validation because
submitting the form with empty data in the required fields would trigger
the error but not an error message.